requirements.txt
Flask==2.3.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.100.0
requests==2.34.0
pytz==2025.8
ffmpeg-python==0.2.0
